create dedicated file for updating db from scraped data and bash script to run it
