+++ kanban-dashboard/update-db.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/bash
+
+ts-node --transpile-only --skip-project ./src/server/api/db.ts
